+++ src/Microsoft.DotNet.Interactive.ExtensionLab/SQLiteConnection.cs
@@ -1,21 +1,25 @@
 ﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System.CommandLine;
 using System.Threading.Tasks;
 using Microsoft.DotNet.Interactive.Connection;
 
 namespace Microsoft.DotNet.Interactive.ExtensionLab
 {
-    public class SQLiteConnection : KernelConnection
+    public class SQLiteConnection : ConnectKernelCommand<SQLiteKernelConnection>
     {
-        public string ConnectionString { get; set; }
-        public override Task<Kernel> ConnectKernelAsync()
+        public SQLiteConnection()
+            : base("sqlite", "Connects to a SQLite database")
         {
-            var kernel = new SQLiteKernel(
-                $"sql-{KernelName}",
-                ConnectionString);
+            Add(new Argument<string>("connectionString", "The connection string used to connect to the database"));
+        }
 
-            return Task.FromResult<Kernel>(kernel);
+        public override Task<Kernel> ConnectKernelAsync(
+            SQLiteKernelConnection kernelConnection,
+            KernelInvocationContext context)
+        {
+            return kernelConnection.ConnectKernelAsync();
         }
     }
 }
