+++ src/Microsoft.DotNet.Interactive/Connection/KernelCommandAndEventReceiverBase.cs
@@ -1,6 +1,6 @@
-﻿
-// Copyright (c) .NET Foundation and contributors. All rights reserved.
+﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
 using System;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
